\docType{class}
\name{Coefficient-class}
\alias{Coefficient-class}
\alias{getCoefficient}
\alias{getCoefficient,Coefficient-method}
\alias{initialize,Coefficient-method}
\title{A coefficient analysis}
\usage{
getCoefficient(object = "Coefficient")
}
\description{
Object of class \code{Coefficient} are created by the
\code{coefAnalysis} functions
}
\details{
An object of the class `Coefficient' has the following
slots: \itemize{ \item \code{coef} A matrix of coefficients
generated by all variable combinations \item \code{Rsquare}
A vector of R-square values from the regressions \item
\code{MeanRsquare} A vector of the mean of R-square values
from the regressions \item \code{x} The first input: A
matrix of covariates \item \code{y} The second input: A
numeric vector length of the number of regressions using
variable combinations }
}
\author{
Elif Ozdemir: \email{eozdemir
}

