\docType{data}
\name{exampleDataset}
\alias{exampleDataset}
\title{Example Dataset}
\description{
The example dataset is a small dataset that illustrates how
the functions work.
}
\details{
The variables included in the dataset are: \itemize{
\item\code{A} A vector of random numbers (independent
variable 1) \item\code{B} Another vector of random numbers
(independent variable 2) \item\code{C} Another vector of
random numbers (independent variable 3) \item\code{D}
Another vector of random numbers (dependent variable) }
}

