\name{plotRegression}
\alias{plotRegression}
\title{Plots the y values and a fits a regression line for Regression class objects}
\usage{
plot(x)
}
\arguments{
  \item{x}{A Regression class object

  \item{x}{The first object input} \item{y}{The second
  object input}}
}
\description{
Plots the y values and a fits a regression line for
Regression class objects
}
\note{
This function only plots for the regression with all
covariates included
}
\author{
Elif Ozdemir: \email{eozdemir
}
\seealso{
\code{\link{regAnalysis}}
}

