\name{printRegression}
\alias{Regression-method}
\alias{printRegression}
\alias{printRegression,}
\title{Prints summary info for Regression class objects}
\usage{
print(x)
}
\arguments{
  \item{x}{A Regression class object}
}
\description{
Prints summary info for Regression class objects
}
\author{
Elif Ozdemir: \email{eozdemir
}
\seealso{
\code{\link{showRegression}}
}

