\docType{class}
\name{Regression-class}
\alias{Regression-class}
\alias{getRegression}
\alias{getRegression,Regression-method}
\alias{initialize,Regression-method}
\title{A regression analysis}
\usage{
getRegression(object = "Regression")
}
\description{
Object of class \code{Regression} are created by the
\code{regAnalysis} functions
}
\details{
An object of the class `Regression' has the following
slots: \itemize{ \item \code{coef} A matrix of coefficients
generated by all variable combinations \item \code{Rsquare}
A vector of R-square values from the regressions \item
\code{x} The first input: A matrix of covariates \item
\code{y} The second input: A numeric vector length of the
number of regressions using variable combinations }
}
\author{
Elif Ozdemir: \email{eozdemir
}

