\name{showRegression}
\alias{showRegression}
\title{Shows some parts of info for Regression class objects}
\usage{
show(x)
}
\arguments{
  \item{x}{A Regression class object}
}
\description{
Shows some parts of info for Regression class objects
}
\note{
Realize that this function works in a different way than
printRegression function does
}
\author{
Elif Ozdemir: \email{eozdemir
}
\seealso{
\code{\link{printRegression}}
}

